refactor: move NewPoolOperation from Pool to PoolV2
