+++ src/templates/PoolV2/NewPoolOperations/Form/InputAndOutputValueToken/index.tsx
