+++ src/templates/PoolV2/NewPoolOperations/Form/InputAndOutputValueToken/styles.ts
