+++ src/templates/PoolV2/NewPoolOperations/Form/ListOfAllAsset/index.tsx
