+++ src/templates/PoolV2/NewPoolOperations/Form/ListOfAllAsset/styles.ts
