+++ src/templates/PoolV2/NewPoolOperations/Form/TokenAssetOut/styles.ts
