+++ src/templates/PoolV2/NewPoolOperations/Form/TokenSelect/styles.ts
