+++ src/templates/PoolV2/NewPoolOperations/Form/TokenSelected/index.tsx
