+++ src/templates/PoolV2/NewPoolOperations/Form/TokenSelection/InputSearch/styles.ts
