+++ src/templates/PoolV2/NewPoolOperations/Form/TokenSelection/index.tsx
