+++ src/templates/PoolV2/NewPoolOperations/Form/Withdraw/styles.ts
