+++ src/templates/PoolV2/NewPoolOperations/SelectOperation/styles.ts
