+++ src/templates/PoolV2/NewPoolOperations/SelectOperationOnMobile/index.tsx
