+++ src/templates/PoolV2/NewPoolOperations/SelectOperationOnMobile/styles.ts
