+++ src/templates/PoolV2/NewPoolOperations/styles.ts
